requirements.txt
streamlit
opencv-python-headless
numpy
scipy
scikit-image
scikit-learn
imbalanced-learn
pillow